Добавил проверку в __set
